remove unused code
I tried to set pool_cls to get the worker to not start another
process. It did work, but the pool_cls cannot be used as kwargs. I did
not find how to start the worker with a thread pool but realized that
xivo-auth worked as expected without starting a worker.