allow persistent tokens to be stored
Split the token manager and the token storage into 2 class. The Storage
class is also used by "xivo-manage-consul" to create tokens that do not
expire. The Storage and Token class might be moved into a separate
package (i.e. out of xivo-auth, and xivo-auth would depends on it),
which would be cleaner.

Also add a notion of a token "name", which is an alternative ID to a
token. This is used to store or update (or get) a given token without
knowing its ID. The ID is a secret information; the name is not. This is
needed by "xivo-manage-consul"; the other solution would have been to
maintain the name -> ID mapping out of consul. The "name" concept is not
exposed by xivo-auth.