verify if plugin should be loaded before load
Event if the plugin is in enabled_plugins, the static method
'should_be_loaded' must return True if you want load the plugin.

We cannot make this verify in the __init__ of plugin because the
plugin will stay in ExtensionManager and if someone try to connect with
the failed plugin, xivo-auth will return a 500.