password reset: add the POST that will set the new password
this route is not decorated by the required_acl decorator to be able to have the same URL to
start and complete the reset password operation.

The POST has a scoped function that is decoration with the required_acl decorator
to be able to inject the user_uuid as a kwargs.