show the contact in the result of a PUT and GET on tenants
The database field is named contact_uuid since its a relation to the auth_user.uuid field.
the marshmallow schema has had a contact field since thats what is in the API but serializing
from the result of a sqlalchemy query resulted in a null contact. This makes the marshmallow
schema use a contact_uuid field and load_from and dump_to "contact"