token: add the metadata object to the token body
the metadata contains information known by wazo-auth that are relevant to this user. This
information can be used to render ACL templates but is not exhaustive as some backends
will optionnaly query other services (xivo-confd).

Since some of those remote queries only happen if the ACL templates uses an ACL
that cannot be rendered without external information. I choose not to add them
to avoid add requirements to systems where xivo-confd is not available and would
not make sense (when using wazo-auth on another system)