external auth: add helpers for oauth2 plugins
Many plugins relying on oauth2 use the following mechanism when creating a new
external auth for a user.

1. Create the some of the user data in the database
2. Start a thread that connects to wss://oauth.wazo.io/ws/<state>
3. Sends a redirect_url to the oauth2 service that will GET on oauth.wazo.io
4. When oauth.wazo.io receives the GET, it writes the parameters of the GET
in the websocket.
5. When wazo-auth receives the data on its websocket, it creates the first token

This requires that all plugins implement the thread handling to wait on the WS
while the user is authorizing wazo-auth in the oauth2 application.

This allow the plugin developer to create only the callback function and
wazo-auth will take care of starting the threads.

The external service can also build the redirect_url on behalf of the plugin.