add the parent_uuid to the tenant table
The goal of this branch is to have ONE tenant on each resource in wazo-auth. Using the parent_uuid
makes the relationship a 1,N.

The migration adds a tenant named "main" that will be the topmost tenant in the hierarchy.

The name "main" should not be used in the wazo-auth code to find the TOP of the tenant hierarchy.
SELECT uuid FROM auth_tenant WHERE auth_tenant.uuid = auth_tenant.parent_uuid
should be used to find the TOP tenant.