tenants: remove cached tenant tree
reason: on the slave, after a DB sync, the webi won't let any user log
in. During the DB sync, the master tenant uuid changes and the cached
tree in wazo-auth is not invalidated. In this case, wazo-auth can't find
the list of tenants of a user and fails at token creation.