debian: decrease debconf questions priority
reason: do not ask the questions when DEBIAN_FRONTEND is not set to
"noninteractive". The questions are quite technical and can be
encountered easily after an upgrade failure. The answers are hard to get
right manually, so rely on the default answers.

By default, questions below "high" priority are never asked. To change
this, use "dpkg-reconfigure debconf".