refresh_token dao: add the count method
the user and tenant segregation are not managed by the filtered flag. A query
with a tenant_uuids list that does not contain the refresh_token's tenant_uuid
will not be able to list it, filtered or not.

the user argument is not filtered because it is part of the URL and that method
is meant to be used in that scenario only.