emails: make the URL more configurable for email templates
the following fields have changed when calling an API that triggers an email

hostname and port are now replaced by the parameters used while doing the
original query

prefix is a new parameter that matches the prefix for this http server

Example:

Given a GET on pcm-dev-0:9497/0.1/....
Then the email hostname is pcm-dev-0 and port 9497

Given a GET on pcm-dev-0/api/auth/0.1/...
Then the email hostname is pcm-dev-0 and prefix is /api/auth

I also modified the email confirmation template which always used the host:port
combination. The email reset template stays the same but the hostname will be
the one used by the reset query instead of being the one configured for service
discovery.