token renewer: commit after creating a new token
Here's the problem.

When the token expires a new one is created. Since the HTTP request is still in
progress the token is not committed to the DB. The token is used to do an
external request to confd which then tries to validate the token. Since the
token has not been commited, auth returns 401 to confd which also returns a 401.