ldap_backend: fix service user bug when user cannot read mail attribute
Explanation: the ldap object used was the same for the service bind and the mail check.
The mail retrieval was done after the user binding was done, so the ldap object was bound
to the final user account. If the user could not retrieve their email address, then it would
not work.