nginx: include cors config explictly
I used [.] in the filename to avoid errors during the installation if the
cors.conf file does not exist on the filesystem at the time of the installation.
Since the brackets make it a regex Nginx will not complain about missing files,
the same way it would behave on an include * inside an empty directory.