startup: fetch group policy association first
adding a policy to a group is idempotent but it does require a roundtrip
to the database. In most cases during startup all groups will already
have the appropriate policy. When there are thousands of tenants this
can take a significant amount of time.

By fetching all group uuid/policy uuid pairs first I was able to remove
~12s from the startup time with 6k tenants and groups going fromk 36s to
24s.