saml: check that the domain exist on load
When loading the SAML configuration from the configuration file we check
that the domain name is configured on a tenant.

This will prevent an administrator from configuring SAML authentication
for a domain that does not exist in wazo-auth. When we will add the SAML
configuration API we will know that a SAML config with an invalid domain
name did not work in the first place and be more confident that we can
ignore it during the migration.