move token extractor to email confirm
why: changing default extractor for the AuthVerifier was confusing
because most of resources use get_tenant_uuids method which hard code
usage of header. Then reading the code make you think query string was
accepted everywhere, but a 401 was raised further in the code

However, in all wazo services, the query string is only accepted on
specific endpoints. So it make more sense to move this check on specific
endpoint