email: smtp logs time without microseconds
this changes cover the case when start_time is in the same second as the
email time.

Since the email has no microseconds it was compared as having a 0
microseconds. Setting the start_time to 0 microseconds will result in
both time being equal when everything happens at the same second